refactor: adjust folder and files names
