+++ src/pages/general/UpdateData/forms/EconomicActivityForm/interface.tsx
@@ -4,16 +4,16 @@ import { Select } from "@design/input/Select";
 import { Grid } from "@design/layout/Grid";
 import { useMediaQuery } from "@hooks/useMediaQuery";
 import { FormikValues } from "formik";
-import { economicActivityDM } from "src/model/domains/economicActivity/economicactivitydme";
-import { contractTypeDM } from "src/model/domains/economicActivity/contracttypedme";
-import { severanceRegimeDM } from "src/model/domains/economicActivity/severanceregimedme";
+import { economicActivityDM } from "src/model/domains/economicActivity/economicactivitydm";
+import { contractTypeDM } from "src/model/domains/economicActivity/contracttypedm";
+import { severanceRegimeDM } from "src/model/domains/economicActivity/severanceregimedm";
 import { workdayDM } from "src/model/domains/economicActivity/workdaydm";
-import { companyFormalityDM } from "src/model/domains/economicActivity/companyformalitydme";
+import { companyFormalityDM } from "src/model/domains/economicActivity/companyformalitydm";
 import { countryDM } from "src/model/domains/financialOperations/countrydm";
 import { ObjectSchema } from "yup";
 import { getFieldState } from "src/utils/forms";
 import { MdSearch } from "react-icons/md";
-import { EconomicActivityModal } from "@components/modals/forms/update-data/EconomicActivityModale";
+import { EconomicActivityModal } from "@components/modals/forms/update-data/EconomicActivityModal";
 import { economicActivitiesMock } from "@mocks/users/economicActivities.mocks";
 import { IEconomicActivity } from "@mocks/users/economicActivities.mocks";
 import { getDomainById } from "@mocks/domains/domainService.mocks";
